clean up

+++ apperception/new_db.py
@@ -1,4 +1,3 @@
-import pickle
 import datetime
 
 import psycopg2
@@ -14,6 +13,8 @@ from video_util import add_recognized_objs, get_video_dimension, recognize
 CAMERA_TABLE = "cameras"
 TRAJ_TABLE = "item_general_trajectory"
 BBOX_TABLE = "general_bbox"
+
+CLEAR_TABLES = True
 
 
 class Database:
@@ -37,31 +38,28 @@ class Database:
         self.start_time = datetime.datetime(2021, 6, 8, 7, 10, 28)
 
     def _create_camera_table(self):
-        # drop old
-        q1 = Query.drop_table(CAMERA_TABLE).if_exists()
+        if CLEAR_TABLES:
+            # drop old
+            q1 = Query.drop_table(CAMERA_TABLE).if_exists()
 
-        # create new
-        q2 = Query.create_table(CAMERA_TABLE).columns(
-            Column("cameraId", "TEXT"),
-            Column("worldId", "TEXT"),
-            Column("ratio", "real"),
-            Column("origin", "geometry"),
-            Column("focalpoints", "geometry"),
-            Column("fov", "INTEGER"),
-            Column("skev_factor", "real"),
-            Column("width", "integer"),
-            Column("height", "integer"),
-        )
+            # create new
+            q2 = Query.create_table(CAMERA_TABLE).columns(
+                Column("cameraId", "TEXT"),
+                Column("worldId", "TEXT"),
+                Column("ratio", "real"),
+                Column("origin", "geometry"),
+                Column("focalpoints", "geometry"),
+                Column("fov", "INTEGER"),
+                Column("skev_factor", "real"),
+                Column("width", "integer"),
+                Column("height", "integer"),
+            )
 
-        self.cur.execute(q1.get_sql())
-        self.cur.execute(q2.get_sql())
-        self.con.commit()
+            self.cur.execute(q1.get_sql())
+            self.cur.execute(q2.get_sql())
+            self.con.commit()
 
-    def insert_cam(self, world_id: str, camera_node_bytes: bytes):
-        camera_node = pickle.loads(camera_node_bytes)
-        if not isinstance(camera_node, Camera):
-            raise Exception('camera_node_bytes should be reconstructed as Camera')
-
+    def insert_cam(self, world_id: str, camera_node: Camera):
         cam = Table(CAMERA_TABLE)
         cam_id = camera_node.cam_id
         cam_ratio = camera_node.ratio
@@ -73,8 +71,8 @@ class Database:
             str(lens.cam_origin[1]),
             str(lens.cam_origin[2]),
         )
-        camera_node.dimension = get_video_dimension(camera_node.video_file)
-        width, height = camera_node.dimension
+        width, height = get_video_dimension(camera_node.video_file)
+        # width, height = camera_node.dimension
 
         q = Query.into(cam).insert(
             cam_id,
@@ -147,14 +145,16 @@ class Database:
         return self.cur.fetchall()
 
     def _create_general_bbox_table(self):
-        # already created in create_or_insert_general_trajectory
-        self.cur.execute("DROP TABLE IF EXISTS General_Bbox;")
-        self.con.commit()
+        if CLEAR_TABLES:
+            # already created in create_or_insert_general_trajectory
+            self.cur.execute("DROP TABLE IF EXISTS General_Bbox;")
+            self.con.commit()
 
     def _create_item_general_trajectory_table(self):
-        # already created in create_or_insert_general_trajectory
-        self.cur.execute("DROP TABLE IF EXISTS Item_General_Trajectory;")
-        self.con.commit()
+        if CLEAR_TABLES:
+            # already created in create_or_insert_general_trajectory
+            self.cur.execute("DROP TABLE IF EXISTS Item_General_Trajectory;")
+            self.con.commit()
 
     def insert_bbox_traj(self, world_id: str, camera_node: Camera, recognition_area: BoundingBox):
         video_file, algo, lens = camera_node.video_file, "Yolo", camera_node.lens
